# analytics-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Analytics Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/analytics/expenses/dates": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Expenses analytics for dates",
                "parameters": [
                    {
                        "enum": [
                            "month",
                            "day"
                        ],
                        "type": "string",
                        "default": "month",
                        "description": "Date unit",
                        "name": "unit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "absolute",
                            "cumulative"
                        ],
                        "type": "string",
                        "default": "absolute",
                        "description": "Calculation type",
                        "name": "calculation",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AnalyticsResponse"
                            }
                        }
                    }
                }
            }
        },
        "/analytics/expenses/groups": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Expenses analytics for groups",
                "parameters": [
                    {
                        "enum": [
                            "category",
                            "merchant"
                        ],
                        "type": "string",
                        "default": "category",
                        "description": "Grouping field",
                        "name": "group",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AnalyticsResponse"
                            }
                        }
                    }
                }
            }
        },
        "/analytics/income/dates": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "income"
                ],
                "summary": "Income analytics for dates",
                "parameters": [
                    {
                        "enum": [
                            "month",
                            "day"
                        ],
                        "type": "string",
                        "default": "month",
                        "description": "Date unit",
                        "name": "unit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AnalyticsResponse"
                            }
                        }
                    }
                }
            }
        },
        "/analytics/income/groups": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "income"
                ],
                "summary": "Income analytics for groups",
                "parameters": [
                    {
                        "enum": [
                            "category",
                            "merchant"
                        ],
                        "type": "string",
                        "default": "category",
                        "description": "Grouping field",
                        "name": "group",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AnalyticsResponse"
                            }
                        }
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Retrieve transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.TransactionResponse"
                            }
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Retrieve transaction details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TransactionResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.AnalyticsGroup": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.AnalyticsResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AnalyticsGroup"
                    }
                }
            }
        },
        "model.TransactionResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "datetime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "merchant": {
                    "type": "string"
                }
            }
        }
    }
}